+++ test/plugin/scenarios/spring-jms-5.x-scenario/configuration.yml
@@ -15,7 +15,7 @@
 # limitations under the License.
 
 type: tomcat
-entryService: http://localhost:8080/spring-jms-5.x-scenario/case/spring-jms-scenario
+entryService: http://localhost:8080/spring-jms-5.x-scenario/case/jms
 healthCheck: http://localhost:8080/spring-jms-5.x-scenario/case/healthCheck
 withPlugins: apm-activemq-5.x-plugin-*.jar
 environment:
