+++ test/plugin/scenarios/spring-jms-5.x-scenario/src/main/webapp/WEB-INF/web.xml
@@ -15,7 +15,6 @@
   ~ limitations under the License.
   ~
   -->
-
 <web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
@@ -24,12 +23,22 @@
     <display-name>skywalking-spring-jms-5.x-scenario</display-name>
 
     <servlet>
-        <servlet-name>spring-mvc</servlet-name>
-        <servlet-class>org.springframework.web.servlet.DispatcherServlet</servlet-class>
-        <load-on-startup>1</load-on-startup>
+        <servlet-name>caseServlet</servlet-name>
+        <servlet-class>org.apache.skywalking.apm.testcase.spring.jms.CaseServlet</servlet-class>
     </servlet>
+
+    <servlet>
+        <servlet-name>healthCheck</servlet-name>
+        <servlet-class>org.apache.skywalking.apm.testcase.spring.jms.HealthCheckServlet</servlet-class>
+    </servlet>
+
     <servlet-mapping>
-        <servlet-name>spring-mvc</servlet-name>
-        <url-pattern>/</url-pattern>
+        <servlet-name>healthCheck</servlet-name>
+        <url-pattern>/case/healthCheck</url-pattern>
+    </servlet-mapping>
+
+    <servlet-mapping>
+        <servlet-name>caseServlet</servlet-name>
+        <url-pattern>/case/jms</url-pattern>
     </servlet-mapping>
 </web-app>
